main/requirements.txt
fastapi==0.115.4
Flask==3.0.3
pyTelegramBotAPI==4.23.0
python-dotenv==1.0.1
Requests==2.32.3
uvicorn==0.32.0
gunicorn==23.0.0